move Core to core. make implicit deps explicit
