+++ CHANGELOG.md
@@ -1,0 +1,1 @@
+See [changelog in docs](docs/changelog.md).
