+++ qcmanybody/__init__.py
@@ -1,7 +1,7 @@
 from importlib.metadata import version
 
-from .manybody import ManyBodyCore
-from .manybody import ManyBodyCore as ManyBodyCalculator  # legacy alias
+from .core import ManyBodyCore
+from .core import ManyBodyCore as ManyBodyCalculator  # legacy alias
 from .computer import ManyBodyComputer
 from .models import BsseEnum
 from .utils import labeler, delabeler, resize_gradient, resize_hessian
