+++ qcmanybody/tests/common.py
@@ -1,4 +1,4 @@
-from qcmanybody.manybody import Molecule
+from qcelemental.models import Molecule
 
 mol_ne2 = Molecule(
     symbols=["Ne", "Ne"], fragments=[[0], [1]], geometry=[0.0, 0.0, -2.834589188186742, 0.0, 0.0, 2.834589188186742]
